"ads2_m12_a1_1: 3/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ ADS-2-assignments/m12/ADS-2_Exam-2/Solution.java
@@ -35,7 +35,7 @@ public class Solution {
             // if ((int) d.distance(Integer.parseInt(paths[1])) == 0) {
             //     System.out.println("No Path Found.");
             // } else {
-                System.out.format("%.1f", (int) d.distance(Integer.parseInt(paths[1])));
+                System.out.format("%.1f",d.distance(Integer.parseInt(paths[1])));
                  break;
 
         case "ViaPaths":
